move upload progress to its own module
